refactor: MyLotto 로또 묶음 함수 리팩토링
